etl/sqlite_to_postgres/requirements.txt
attrs==22.2.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pytest==7.2.2
python-dotenv==1.0.0
